add debeziium-supplier-properties for the supplier. docs streamlining
