+++ applications/source/debezium-source/src/main/resources/META-INF/dataflow-configuration-metadata-whitelist.properties
@@ -1,1 +1,1 @@
-configuration-properties.classes=org.springframework.cloud.fn.common.debezium.DebeziumProperties
+configuration-properties.classes=org.springframework.cloud.fn.common.debezium.DebeziumProperties,org.springframework.cloud.fn.supplier.debezium.DebeziumSupplierProperties
