+++ functions/supplier/debezium-supplier/pom.xml
@@ -104,21 +104,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-        <!-- <dependency>
-            <groupId>io.debezium</groupId>
-            <artifactId>debezium-connector-cassandra</artifactId>
-            <exclusions>
-                <exclusion>
-                    <artifactId>slf4j-log4j12</artifactId>
-                    <groupId>org.slf4j</groupId>
-                </exclusion>
-                <exclusion>
-                    <artifactId>hibernate-validator</artifactId>
-                    <groupId>org.hibernate</groupId>
-                </exclusion>
-            </exclusions>
-            <optional>true</optional>
-        </dependency> -->
         <dependency>
             <groupId>io.debezium</groupId>
             <artifactId>debezium-connector-spanner</artifactId>
@@ -130,18 +115,6 @@
                 </exclusion>
             </exclusions>
             <optional>true</optional>
-        </dependency>
-
-        <!-- APICURIO Schema Registry (for AVRO format)-->
-        <dependency>
-            <groupId>io.apicurio</groupId>
-            <artifactId>apicurio-registry-utils-converter</artifactId>
-            <version>${apicurio.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.apicurio</groupId>
-            <artifactId>apicurio-registry-client</artifactId>
-            <version>${apicurio.version}</version>
         </dependency>
 
         <!-- TEST -->
